requirements.txt
requests>=2.31.0,<3.0.0
google-adk>=0.1.0,<0.2.0
pydantic>=2.11.3,<3.0.0
python-dotenv>=1.1.0,<2.0.0
google-cloud-aiplatform[adk,agent_engines]>=1.59.0,<2.0.0
absl-py>=2.1.0,<3.0.0
cloudpickle>=3.0.0,<4.0.0
fire==0.6.0